# api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.2 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create sounds
	// (POST /admin/sounds)
	CreateSounds(ctx echo.Context) error
	// Get playlists
	// (GET /playlists)
	GetPlaylists(ctx echo.Context) error
	// Create playlists
	// (POST /playlists)
	CreatePlaylists(ctx echo.Context) error
	// Get sounds
	// (GET /sounds)
	GetSounds(ctx echo.Context) error
	// Get recommended sounds
	// (GET /sounds/recommended)
	GetRecommendedSounds(ctx echo.Context, params GetRecommendedSoundsParams) error
	// Get sound
	// (GET /sounds/{id})
	GetSound(ctx echo.Context, id string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateSounds converts echo context to params.
func (w *ServerInterfaceWrapper) CreateSounds(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateSounds(ctx)
	return err
}

// GetPlaylists converts echo context to params.
func (w *ServerInterfaceWrapper) GetPlaylists(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetPlaylists(ctx)
	return err
}

// CreatePlaylists converts echo context to params.
func (w *ServerInterfaceWrapper) CreatePlaylists(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreatePlaylists(ctx)
	return err
}

// GetSounds converts echo context to params.
func (w *ServerInterfaceWrapper) GetSounds(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSounds(ctx)
	return err
}

// GetRecommendedSounds converts echo context to params.
func (w *ServerInterfaceWrapper) GetRecommendedSounds(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRecommendedSoundsParams
	// ------------- Optional query parameter "playlistId" -------------

	err = runtime.BindQueryParameter("form", true, false, "playlistId", ctx.QueryParams(), &params.PlaylistId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter playlistId: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetRecommendedSounds(ctx, params)
	return err
}

// GetSound converts echo context to params.
func (w *ServerInterfaceWrapper) GetSound(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSound(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/admin/sounds", wrapper.CreateSounds)
	router.GET(baseURL+"/playlists", wrapper.GetPlaylists)
	router.POST(baseURL+"/playlists", wrapper.CreatePlaylists)
	router.GET(baseURL+"/sounds", wrapper.GetSounds)
	router.GET(baseURL+"/sounds/recommended", wrapper.GetRecommendedSounds)
	router.GET(baseURL+"/sounds/:id", wrapper.GetSound)

}