# api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.2 DO NOT EDIT.
package api

import (
	"time"
)

// Credit defines model for Credit.
type Credit struct {
	Name string `json:"name"`
	Role string `json:"role"`
}

// NewPlaylist defines model for NewPlaylist.
type NewPlaylist struct {
	Sounds []string `json:"sounds"`
	Title  string   `json:"title"`
}

// NewPlaylistRequest defines model for NewPlaylistRequest.
type NewPlaylistRequest struct {
	Data []NewPlaylist `json:"data"`
}

// NewSound defines model for NewSound.
type NewSound struct {
	Bpm               int32    `json:"bpm"`
	Credits           []Credit `json:"credits"`
	DurationInSeconds int32    `json:"duration_in_seconds"`
	Genres            []string `json:"genres"`
	Title             string   `json:"title"`
}

// NewSoundRequest defines model for NewSoundRequest.
type NewSoundRequest struct {
	Data []NewSound `json:"data"`
}

// Playlist defines model for Playlist.
type Playlist struct {
	Id     string   `json:"id"`
	Sounds []string `json:"sounds"`
	Title  string   `json:"title"`
}

// PlaylistResponse defines model for PlaylistResponse.
type PlaylistResponse struct {
	Data []Playlist `json:"data"`
}

// Sound defines model for Sound.
type Sound struct {
	Bpm               int32     `json:"bpm"`
	CreatedAt         time.Time `json:"created_at"`
	Credits           []Credit  `json:"credits"`
	DurationInSeconds int32     `json:"duration_in_seconds"`
	Genres            []string  `json:"genres"`
	Id                string    `json:"id"`
	Title             string    `json:"title"`
	UpdatedAt         time.Time `json:"updated_at"`
}

// SoundResponse defines model for SoundResponse.
type SoundResponse struct {
	Data []Sound `json:"data"`
}

// GetRecommendedSoundsParams defines parameters for GetRecommendedSounds.
type GetRecommendedSoundsParams struct {
	PlaylistId *string `form:"playlistId,omitempty" json:"playlistId,omitempty"`
}

// CreateSoundsJSONRequestBody defines body for CreateSounds for application/json ContentType.
type CreateSoundsJSONRequestBody = NewSoundRequest

// CreatePlaylistsJSONRequestBody defines body for CreatePlaylists for application/json ContentType.
type CreatePlaylistsJSONRequestBody = NewPlaylistRequest